updated main.c for ep3
this update was done on episode 3 of the YouTube series and in this I added the includes and some typedefs also a place holder for a file reader function has been added. the next update is on episode 4 where i will be adding the parts from the to-do list in the comments.